Split Resque tests per module
Don't test everything in the hooks spec, but test what's done per spec
file.

[skip changeset]
[skip review]
